requirements.txt
flask
pytesseract
pdfplumber
Pillow
requests

